+++ scrapers/treehugger.py
@@ -1,31 +1,9 @@
 from bs4 import BeautifulSoup
 import requests
+from utils import * 
 
 ## for treehugger.com
 ## does not include featured article, can include all articles once duplicate catcher is in place?
- 
-months = {
-    'January': 1,
-    'February': 2,
-    'March': 3,
-    'April': 4,
-    'May': 5,
-    'June': 6,
-    'July': 7,
-    'August': 8,
-    'September': 9,
-    'October': 10,
-    'November': 11,
-    'December' :12
-}
-
-def date_convert(string):
-    date = string.split()
-    mon = months[date[0]]
-    day = date[1].strip(',')
-    year = date[2]
-    return "{}/{}/{}".format(mon, day, year)
-
 
 def scrape_treehugger():
     '''
@@ -71,5 +49,5 @@ def scrape_treehugger():
 
 
 if __name__ == "__main__":
-    print(treehugger())
+    print(scrape_treehugger())
