+++ scrapers/utils.py
@@ -1,0 +1,21 @@
+months = {
+    'January': 1,
+    'February': 2,
+    'March': 3,
+    'April': 4,
+    'May': 5,
+    'June': 6,
+    'July': 7,
+    'August': 8,
+    'September': 9,
+    'October': 10,
+    'November': 11,
+    'December' :12
+}
+
+def date_convert(string):
+    date = string.split()
+    mon = months[date[0]]
+    day = date[1].strip(',')
+    year = date[2]
+    return "{}/{}/{}".format(mon, day, year)
